v2.0.10
- New version
- viu auto install processing added
